refactor(whatsapp): move instance management to integrations page
Move WhatsApp Instance Management from Account Settings to the
Integrations page for better discoverability and consistency with
other integrations like Slack, Linear, and Notion.

Changes:
- Add whatsapp_web integration entry to apps.yml
- Create WhatsappWeb/Index.vue component in integrations
- Add route for /integrations/whatsapp_web
- Update Vuex store to compute enabled status from account settings
- Add i18n strings to integrations.json and en.yml
- Remove WhatsApp section from Account Settings page
- Update feature plan documentation
